Add ProductItems as a nested blueprint under products
Allows create and update. Here I split the attributes between Product
and ProductItem such that Product is a category, always having one or
more ProductItem. In the case of fungible items, the ProductItem is
conceptually a variant of the product. In the case of non-fungible
items, each ProductItem represents a separate item, and has a
amount_in_stock value of 1.